go-fundamentals/roman-numerals: An intro to property based tests

+++ go-fundamentals/roman-numerals/numeral.go
@@ -3,7 +3,7 @@ package main
 import "strings"
 
 type RomanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
@@ -25,7 +25,7 @@ var allRomanNumerals = []RomanNumeral{
 	{1, "I"},
 }
 
-func ConvertToRoman(arabic int) string {
+func ConvertToRoman(arabic uint16) string {
 
 	var result strings.Builder
 
@@ -39,8 +39,8 @@ func ConvertToRoman(arabic int) string {
 	return result.String()
 }
 
-func ConvertToArabic(roman string) int {
-	var arabic = 0
+func ConvertToArabic(roman string) uint16 {
+	var arabic = uint16(0)
 
 	for _, numeral := range allRomanNumerals {
 		for strings.HasPrefix(roman, numeral.Symbol) {
@@ -52,7 +52,7 @@ func ConvertToArabic(roman string) int {
 	return arabic
 }
 
-func (r RomanNumerals) ValueOf(symbol string) int {
+func (r RomanNumerals) ValueOf(symbol string) uint16 {
 	for _, s := range r {
 		if s.Symbol == symbol {
 			return s.Value
